requirements.txt
Flask~=2.3.2
Flask-SQLAlchemy
Flask-Migrate~=3.1.0
psycopg2-binary
Werkzeug~=2.3.4
pytest
python-dotenv~=1.0.1

PyJWT~=2.10.1
alembic~=1.14.0
SQLAlchemy~=2.0.36
jwt~=1.3.1
Flask-Testing~=0.8.1